% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-trie.R
\name{as_asntrie}
\alias{as_asntrie}
\title{Convert a `routeviews` data frame to a trie}
\usage{
as_asntrie(x, ...)
}
\arguments{
\item{x}{a `routeviews` data frame read in with [routeviews_current()] or
[parse_routeviews()]}

\item{...}{ignored (for now)}
}
\value{
an `routeviews_asntrie`
}
\description{
Convert a `routeviews` data frame to a trie
}
\examples{
\dontrun{
rv_df <- routeviews_latest()
rv_trie <- as_asntrie(rv_df)
iptools::ip_to_asn("174.62.167.97", rv_trie)
}
}
