% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asnames-current.R
\name{asnames_current}
\alias{asnames_current}
\title{Retrive current \code{asnames.txt} from \code{bgp.poraroo.net}}
\usage{
asnames_current()
}
\value{
data frame (tibble) with "\code{asn}", "\code{handle}", "\code{asinfo}", "\code{iso2c}"
}
\description{
Retrive current \code{asnames.txt} from \code{bgp.poraroo.net}
}
\note{
This function creates "\code{~/.asnames}" and stores a timestamped copy of
the file in it. These are ~2MB in size so you should peridically
clean out "\code{~/.asnames}" to reclaim disk space.
}
\examples{
\dontrun{
asnames_current()
}
}
\references{
\url{https://bgp.potaroo.net/as6447/asnames.txt}
}
\seealso{
read_asnames

parse_asnames
}
