% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asnames-parse.R
\name{parse_asnames}
\alias{parse_asnames}
\title{Parse arbitrary \code{asnames.txt} downloaded from \code{bgp.poraroo.net}}
\usage{
parse_asnames(as_fil)
}
\arguments{
\item{as_fil}{path to an \code{asnames.txt} file. \code{\link[=path.expand]{path.expand()}} will be run on this value}
}
\value{
data frame (tibble) with "\code{asn}", "\code{handle}", "\code{asinfo}", "\code{iso2c}"
}
\description{
If you don't want \code{\link[=asnames_current]{asnames_current()}} and [read_asnames())] to manage
\code{asnames.txt} files for you, you can use this function to parse \code{asnames.txt}
files you've downloaded and manage via other means.

[read_asnames())]: R:read_asnames())
}
\examples{
\dontrun{
   parse_asnames("/tmp/asnames.txt")
}
}
\references{
\url{https://bgp.potaroo.net/as6447/asnames.txt}
}
\seealso{
asnames_current

read_asnames
}
