% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-asorg2info.R
\name{parse_asorg2info}
\alias{parse_asorg2info}
\title{Parses CAIDA \code{asorg2info} Files}
\usage{
parse_asorg2info(as_org2info_fil)
}
\arguments{
\item{as_org2info_fil}{path to an "org2info" file (e.g. \code{20181001.as-org2info.txt.gz});
\code{\link[=path.expand]{path.expand()}} will be run on this value.}
}
\value{
list with two data frames (tibbles): \code{asorg_info} containing
"\code{org_id}", "\code{changed}", "\code{org_name}", "\code{country}", "\code{source}"; and,
\code{aut_info} containing "\code{aut}", "\code{changed}", "\code{aut_name}", "\code{org_id}",
"\code{opaque_id}", "\code{source}"
}
\description{
Please refer to \url{http://data.caida.org/datasets/as-organizations/} for
the data files, acceptable use information and citation information.
}
\section{AS fields}{


\code{aut} : the AS number
\code{changed} : the changed date provided by its WHOIS entry
\code{aut_name} : the name provide for the individual AS number
\code{org_id} : maps to an organization entry
\code{opaque_id} : opaque identifier used by RIR extended delegation format
\code{source} : the RIR or NIR database which was contained this entry
}

\section{Organization fields}{

\itemize{
\item \code{org_id} : unique ID for the given organization
\item \code{changed} : the changed date provided by its WHOIS entry
\item \code{name} : name could be selected from the AUT entry tied to the organization, the AUT entry with the largest customer cone, listed for the organization (if there existed an stand alone organization), or a human maintained file.
\item \code{country} : some WHOIS provide as a individual field; inferred if not provided
\item \code{source} : the RIR or NIR database which was contained this entry
}
}

\examples{
\dontrun{
parse_asorg2info("20181001.as-org2info.txt.gz")
}
}
\references{
\url{http://data.caida.org/datasets/as-organizations/}
}
