% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-routeviews.R
\name{parse_routeviews}
\alias{parse_routeviews}
\title{Parses Arbitrary RouteViews "Prefix-to-AS" Files}
\usage{
parse_routeviews(rv_fil)
}
\arguments{
\item{rv_fil}{path to a CAIDA Prefix-to-AS file. This will be processed with \code{\link[=path.expand]{path.expand()}}.}
}
\value{
a data frame (tibble) with "\code{cidr}" and "\code{asn}" columns.
}
\description{
CAIDA maintains current & historical IPv4/IPv6 Prefix-to-Autonomous System (AS)
mappings that have been derived from \href{http://www.routeviews.org}{RouteViews} data.
}
\details{
See \href{http://data.caida.org/datasets/routing/routeviews-prefix2as>}{the datasets repository}
for usage and citation information.
}
\examples{
\dontrun{
# you should cache this vs directly pull OR cache after pulling to avoid
# wasting CAIDA bandwidth

parse_routeviews("http://data.caida.org/datasets/routing/routeviews-prefix2as/2018/12/routeviews-rv2-20181203-1200.pfx2as.gz")
}
}
\references{
\url{http://data.caida.org/datasets/routing/routeviews-prefix2as}
}
\seealso{
routeviews_latest
}
