% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asnames-read.R
\name{read_asnames}
\alias{read_asnames}
\title{Retrieve an historical \code{asnames.txt} from local cache}
\usage{
read_asnames(date = Sys.Date())
}
\arguments{
\item{date}{character (ISO date) or \code{Date} of \code{asnames.txt} to read from "\code{~/.asnames}";}
}
\value{
data frame (tibble) with "\code{asn}", "\code{handle}", "\code{asinfo}", "\code{iso2c}"
}
\description{
Looks for and loads a \code{date} timestamped \code{asnames.txt} from "\code{~/.asnames}".
}
\details{
You can see available \code{asnames.txt} files via \code{list.files("~/.asnames")} and
read/parse arbitrary \code{asnames.txt} files with \code{\link[=parse_asnames]{parse_asnames()}}.
}
\examples{
\dontrun{
   read_asname("2018-12-05")
}
}
\references{
\url{https://bgp.potaroo.net/as6447/asnames.txt}
}
\seealso{
asnames_current

parse_asnames
}
