% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routeviews-latest.R
\name{routeviews_latest}
\alias{routeviews_latest}
\title{Caches & Returns  the latest CAIDA-processed RouteViews "Prefix-to-AS" File}
\usage{
routeviews_latest()
}
\value{
a data frame (tibble) with "\code{cidr}" and "\code{asn}" columns.
}
\description{
CAIDA maintains current & historical IPv4/IPv6 Prefix-to-Autonomous System (AS)
mappings that have been derived from \href{http://www.routeviews.org}{RouteViews} data.
This function figures the most current file, caches it and returns a data frame.
}
\details{
You can see available, historical files via \code{list.files("~/.routeviews")} and
read/parse arbitrary prefix-to-AS CAIDA RouteViews files with \code{\link[=parse_routeviews]{parse_routeviews()}}.

See \href{http://data.caida.org/datasets/routing/routeviews-prefix2as>}{the datasets repository}
for usage and citation information.
}
\note{
A "\code{~/.routeviews}" directory will be created if it does not exist.
}
\examples{
\dontrun{
routeviews_latest()
}
}
\references{
\url{http://data.caida.org/datasets/routing/routeviews-prefix2as}
}
\seealso{
parse_routeviews
}
