% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astools-package.R
\docType{package}
\name{astools}
\alias{astools}
\alias{astools-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/astools}
\item BugReports: \url{https://gitlab.com/hrbrmstr/astools/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
